requirements.txt
numpy~=1.23.5
pandas~=1.5.3
scipy~=1.10.1
pyyaml~=6.0
tqdm~=4.65.0
torchdiffeq~=0.2.3
torchsde~=0.2.6
matplotlib~=3.7.1
fastdtw~=0.3.4
statsmodels~=0.14.0